Refactor : 콘솔로그 제거 및 임포트 경로 수정
